+++ src/ActivityScene.jsx
@@ -3,11 +3,19 @@ import { Output } from './Output';
 import { CommandForm } from './CommandForm';
 
 export class ActivityScene extends React.Component {
+	renderChildren() {
+		return (
+			<>
+				<Output {...this.props} />
+				<CommandForm prompt={this.props.state ? this.props.state.prompt : ''} />
+			</>
+		);
+	}
+
 	render() {
 		return (
 			<div className="ActivityScene">
-				<Output {...this.props} />
-				<CommandForm prompt={this.props.state ? this.props.state.prompt : ''} />
+				{ this.props.children || this.renderChildren() }
 			</div>
 		);
 	}