+++ src/ConclusionScene.jsx
@@ -2,10 +2,14 @@ import React from 'react';
 import {Output} from './Output';
 
 export class ConclusionScene extends React.Component {
+	renderChildren() {
+		return <Output {...this.props} />;
+	}
+
 	render() {
 		return (
 			<div className="ConclusionScene">
-				<Output {...this.props} />
+				{ this.props.children || this.renderChildren() }
 			</div>
 		);
 	}