+++ test/MultipleChoiceScene.test.js
@@ -3,12 +3,12 @@ import { configure, shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
 import { MultipleChoiceScene } from '../src/MultipleChoiceScene';
-import { CommandLink } from '../src/CommandLink';
+import { ChoiceList } from '../src/ChoiceList';
 
 configure({ adapter: new Adapter() });
 
 describe('<MultipleChoiceScene />', () => {
-    it('renders options', () => {
+    it('renders a ChoiceList', () => {
         let props = {
             state: {
                 options: ["First", "Second"],
@@ -16,6 +16,11 @@ describe('<MultipleChoiceScene />', () => {
             }
         }
         let scene = shallow(<MultipleChoiceScene {...props} />);
-        expect(scene.find(CommandLink).length).toBe(2);
+        expect(scene.find(ChoiceList).length).toBe(1);
     })
+
+    it('renders children', () => {
+        let scene = shallow(<MultipleChoiceScene><img /></MultipleChoiceScene>);
+        expect(scene.find('img').length).toBe(1);
+    });
 });